machina/pyproject.toml
[project]
name = "machina"
version = "0.1.0"
description = "MachinaOS project supervisor CLI"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "rich>=13",
  "psutil>=6",
  "anyio>=4",
  "python-dotenv>=1",
  "pywin32>=308; platform_system == 'Windows'",
]

[project.scripts]
machina = "machina.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["machina*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
